Предварительно разложил классы по папкам
