--- app/src/main/java/com/example/playlistmaker/Creator.kt
@@ -1,17 +1,0 @@
-package com.example.playlistmaker
-
-import android.media.MediaPlayer
-import com.example.playlistmaker.data.PlayerRepositoryImpl
-import com.example.playlistmaker.domain.api.PlayerInteractor
-import com.example.playlistmaker.domain.api.PlayerRepository
-import com.example.playlistmaker.domain.impl.PlayerInteractorImpl
-
-object Creator {
-    fun getRepository(): PlayerRepository {
-        return PlayerRepositoryImpl(MediaPlayer())
-    }
-
-    fun providePlayerInteractor(): PlayerInteractor {
-        return PlayerInteractorImpl(getRepository())
-    }
-}
