+++ app/src/main/java/com/example/playlistmaker/creator/Creator.kt
@@ -1,0 +1,17 @@
+package com.example.playlistmaker.creator
+
+import android.media.MediaPlayer
+import com.example.playlistmaker.player.data.network.PlayerRepositoryImpl
+import com.example.playlistmaker.player.domain.api.PlayerInteractor
+import com.example.playlistmaker.player.domain.api.PlayerRepository
+import com.example.playlistmaker.player.domain.impl.PlayerInteractorImpl
+
+object Creator {
+    fun getRepository(): PlayerRepository {
+        return PlayerRepositoryImpl(MediaPlayer())
+    }
+
+    fun providePlayerInteractor(): PlayerInteractor {
+        return PlayerInteractorImpl(getRepository())
+    }
+}
