--- app/src/main/java/com/example/playlistmaker/data/dto/TrackResponse.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-import com.example.playlistmaker.domain.models.Track
-
-class TrackResponse(
-    val resultCount: Int,
-    val results: ArrayList<Track>) {
-}
