--- app/src/main/java/com/example/playlistmaker/data/local/SearchHistory.kt
@@ -1,10 +1,0 @@
-package com.example.playlistmaker.data.local
-
-import android.content.SharedPreferences
-import com.example.playlistmaker.domain.models.Track
-
-interface SearchHistory {
-    fun read(): List<Track>
-    fun write(track: Track)
-    fun clear()
-}
