--- app/src/main/java/com/example/playlistmaker/data/local/TrackMapper.kt
@@ -1,38 +1,0 @@
-package com.example.playlistmaker.data.local
-
-import com.example.playlistmaker.data.dto.TrackDto
-import com.example.playlistmaker.domain.models.Track
-
-class TrackMapper {
-
-    fun mapToDomain(trackDto: TrackDto): Track {
-        return Track(
-            trackName = trackDto.trackName,
-            artistName = trackDto.artistName,
-            trackTimeMillis = trackDto.trackTimeMillis,
-            artworkUrl100 = trackDto.artworkUrl100,
-            trackId = trackDto.trackId,
-            collectionName = trackDto.collectionName,
-            releaseDate = trackDto.releaseDate,
-            primaryGenreName = trackDto.primaryGenreName,
-            country = trackDto.country,
-            previewUrl = trackDto.previewUrl,
-        )
-    }
-
-    fun mapToDto(track: Track): TrackDto {
-        return TrackDto(
-            trackName = track.trackName,
-            artistName = track.artistName,
-            trackTimeMillis = track.trackTimeMillis,
-            artworkUrl100 = track.artworkUrl100,
-            trackId = track.trackId,
-            collectionName = track.collectionName,
-            releaseDate = track.releaseDate,
-            primaryGenreName = track.primaryGenreName,
-            country = track.country,
-            previewUrl = track.previewUrl,
-        )
-    }
-
-}
