--- app/src/main/java/com/example/playlistmaker/domain/models/PlayerState.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.domain.models
-
-enum class PlayerState {
-    STATE_PREPARED ,
-    STATE_PLAYING,
-    STATE_PAUSED,
-    STATE_COMPLETE
-}
