+++ app/src/main/java/com/example/playlistmaker/player/data/dto/TrackDto.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.player.data.dto
+
+data class TrackDto (
+    val trackName: String,      // Название композиции
+    val artistName: String,     // Имя исполнителя
+    val trackTimeMillis: Long,     // Продолжительность трека
+    val artworkUrl100: String?,    // ссылка на мини-обложку
+    val trackId: Long?, // Ссылка на изображение обложки
+    val collectionName: String?,    // название альбома
+    val releaseDate: String?, // Год выпуска
+    val primaryGenreName: String, // Жанр
+    val country: String,  // страна выпуска трека
+    val previewUrl: String, // ссылка на отрывок песни в 30 секунд
+)
