+++ app/src/main/java/com/example/playlistmaker/player/data/local/SearchHistory.kt
@@ -1,0 +1,9 @@
+package com.example.playlistmaker.player.data.local
+
+import com.example.playlistmaker.player.domain.models.Track
+
+interface SearchHistory {
+    fun read(): List<Track>
+    fun write(track: Track)
+    fun clear()
+}
