+++ app/src/main/java/com/example/playlistmaker/player/data/network/PlayerRepositoryImpl.kt
@@ -1,0 +1,40 @@
+package com.example.playlistmaker.player.data.network
+
+import android.media.MediaPlayer
+import com.example.playlistmaker.player.domain.api.PlayerRepository
+import com.example.playlistmaker.player.domain.models.PlayerState
+
+class PlayerRepositoryImpl(private val mediaPlayer: MediaPlayer): PlayerRepository {
+
+    private var stateCallback: ((PlayerState) -> Unit)? = null
+    override fun preparePlayer(url: String) {
+        mediaPlayer.setDataSource(url)
+        mediaPlayer.prepareAsync()
+        mediaPlayer.setOnPreparedListener {
+            stateCallback?.invoke(PlayerState.STATE_PREPARED)
+        }
+        mediaPlayer.setOnCompletionListener {
+            stateCallback?.invoke(PlayerState.STATE_COMPLETE)
+        }
+    }
+
+    override fun startPlayer() {
+        mediaPlayer.start()
+        stateCallback?.invoke(PlayerState.STATE_PLAYING)
+    }
+
+    override fun pausePlayer() {
+        mediaPlayer.pause()
+        stateCallback?.invoke(PlayerState.STATE_PAUSED)
+    }
+
+    override fun getPosition() = mediaPlayer.currentPosition.toLong()
+
+    override fun setOnStateChangeListener(callback: (PlayerState) -> Unit) {
+        stateCallback = callback
+    }
+
+    override fun reset() {
+        mediaPlayer.reset()
+    }
+}
