+++ app/src/main/java/com/example/playlistmaker/player/domain/api/PlayerInteractor.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.player.domain.api
+
+import com.example.playlistmaker.player.domain.models.PlayerState
+
+interface PlayerInteractor {
+    fun preparePlayer(url: String)
+    fun startPlayer()
+    fun pausePlayer()
+    fun setOnStateChangeListener(callback: (PlayerState) -> Unit)
+
+    fun getPosition(): Long
+    fun reset()
+
+}
