+++ app/src/main/java/com/example/playlistmaker/player/domain/impl/PlayerInteractorImpl.kt
@@ -1,0 +1,31 @@
+package com.example.playlistmaker.player.domain.impl
+
+import com.example.playlistmaker.player.domain.api.PlayerInteractor
+import com.example.playlistmaker.player.domain.api.PlayerRepository
+import com.example.playlistmaker.player.domain.models.PlayerState
+
+class PlayerInteractorImpl(private val repository: PlayerRepository): PlayerInteractor {
+    override fun preparePlayer(url: String) {
+        repository.preparePlayer(url)
+    }
+
+    override fun startPlayer() {
+        repository.startPlayer()
+    }
+
+    override fun pausePlayer() {
+        repository.pausePlayer()
+    }
+
+    override fun setOnStateChangeListener(callback: (PlayerState) -> Unit) {
+        repository.setOnStateChangeListener(callback)
+    }
+
+    override fun getPosition(): Long {
+        return repository.getPosition()
+    }
+
+    override fun reset() {
+        repository.reset()
+    }
+}
