+++ app/src/main/java/com/example/playlistmaker/player/domain/models/PlayerState.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.player.domain.models
+
+enum class PlayerState {
+    STATE_PREPARED ,
+    STATE_PLAYING,
+    STATE_PAUSED,
+    STATE_COMPLETE
+}
