+++ app/src/main/java/com/example/playlistmaker/player/ui/PlayerViewModel.kt
@@ -1,0 +1,63 @@
+package com.example.playlistmaker.player.ui
+
+import android.icu.text.SimpleDateFormat
+import androidx.lifecycle.ViewModel
+
+import android.os.Handler
+import android.os.Looper
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import com.example.playlistmaker.player.domain.api.PlayerInteractor
+import com.example.playlistmaker.player.domain.models.PlayerState
+import com.example.playlistmaker.player.domain.models.Track
+import java.util.Locale
+
+class PlayerViewModel(
+    private val track: Track,
+    private val playerInteractor: PlayerInteractor,
+) : ViewModel() {
+
+    init {
+        playerInteractor.preparePlayer(track.previewUrl)
+
+        playerInteractor.setOnStateChangeListener { state ->
+            stateLiveData.postValue(state)
+            if (state == PlayerState.STATE_COMPLETE) handler.removeCallbacks(time)
+        }
+    }
+
+    private val handler = Handler(Looper.getMainLooper())
+
+    private val time = object : Runnable {
+        override fun run() {
+            val position = playerInteractor.getPosition()
+            timeLiveData.postValue(SimpleDateFormat("mm:ss", Locale.getDefault()).format(position))
+            handler.postDelayed(this, TRACK_TIME_DELAY)
+        }
+    }
+
+    private val stateLiveData = MutableLiveData<PlayerState>()
+    fun observeState(): LiveData<PlayerState> = stateLiveData
+
+    private val timeLiveData = MutableLiveData<String>()
+    fun observeTime(): LiveData<String> = timeLiveData
+
+    fun play() {
+        playerInteractor.startPlayer()
+        handler.post(time)
+    }
+
+    fun pause() {
+        playerInteractor.pausePlayer()
+        handler.removeCallbacks(time)
+    }
+
+    fun release() {
+        playerInteractor.reset()
+        handler.removeCallbacks(time)
+    }
+
+    companion object {
+        private const val TRACK_TIME_DELAY = 300L
+    }
+}
