+++ app/src/main/java/com/example/playlistmaker/search/ui/TrackListViewHolder.kt
@@ -1,0 +1,34 @@
+package com.example.playlistmaker.search.ui
+
+import android.icu.text.SimpleDateFormat
+import android.view.View
+import android.widget.ImageView
+import android.widget.LinearLayout
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.playlistmaker.R
+import com.example.playlistmaker.player.domain.models.Track
+import java.util.*
+
+class TrackListViewHolder(itemView: View): RecyclerView.ViewHolder(itemView) {
+    private val rootLayout: LinearLayout = itemView.findViewById(R.id.trackRootLayout)
+    private val trackCover: ImageView = itemView.findViewById(R.id.trackCover)
+    private val trackName: TextView = itemView.findViewById(R.id.trackName)
+    private val artistName: TextView = itemView.findViewById(R.id.artistName)
+    private val trackTime: TextView = itemView.findViewById(R.id.trackTime)
+
+
+    fun bind(item: Track) {
+        trackName.text = item.trackName
+        artistName.text = item.artistName
+        trackTime.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(item.trackTimeMillis)
+        Glide.with(trackCover)
+            .load(item.artworkUrl100)
+            .transform(RoundedCorners(itemView.resources.getDimensionPixelSize(R.dimen.cover_corner_radius)))
+            .placeholder(R.drawable.albumcover_placeholder)
+            .into(trackCover)
+    }
+
+}
